+++ CopilotKit/packages/backend/src/lib/langchain-adapter.ts
@@ -3,13 +3,16 @@ import {
   AIMessage,
   BaseMessage,
   BaseMessageChunk,
+  FunctionMessage,
   HumanMessage,
   SystemMessage,
+  ToolMessage,
 } from "@langchain/core/messages";
 import { IterableReadableStream } from "@langchain/core/utils/stream";
 import { CopilotKitServiceAdapter } from "../types";
 import { writeChatCompletionChunk, writeChatCompletionEnd } from "../utils";
 import { CopilotKitResponse } from "../types/service-adapter";
+import { SingleChunkReadableStream } from "../utils";
 
 export type LangChainMessageStream = IterableReadableStream<BaseMessageChunk>;
 export type LangChainReturnType = LangChainMessageStream | BaseMessageChunk | string | AIMessage;
@@ -78,13 +81,37 @@ export class LangChainAdapter implements CopilotKitServiceAdapter {
     // map messages to langchain format
     if (forwardedProps.messages && Array.isArray(forwardedProps.messages)) {
       const newMessages: BaseMessage[] = [];
+
       for (const message of forwardedProps.messages) {
         if (message.role === "user") {
           newMessages.push(new HumanMessage(message.content));
         } else if (message.role === "assistant") {
-          newMessages.push(new AIMessage(message.content));
+          if (message.function_call) {
+            newMessages.push(
+              new AIMessage({
+                content: "",
+                tool_calls: [
+                  {
+                    id: message.function_call.name + "-" + forwardedProps.messages.indexOf(message),
+                    args: JSON.parse(message.function_call.arguments),
+                    name: message.function_call.name,
+                  },
+                ],
+              }),
+            );
+          } else {
+            newMessages.push(new AIMessage(message.content));
+          }
         } else if (message.role === "system") {
           newMessages.push(new SystemMessage(message.content));
+        } else if (message.role == "function") {
+          // An assistant message with 'tool_calls' must be followed by tool messages responding to each 'tool_call_id'
+          newMessages.push(
+            new ToolMessage({
+              content: message.content,
+              tool_call_id: message.name + "-" + (forwardedProps.messages.indexOf(message) - 1),
+            }),
+          );
         }
       }
       forwardedPropsCopy.messages = newMessages;
@@ -151,34 +178,6 @@ export class LangChainAdapter implements CopilotKitServiceAdapter {
       cancel() {
         cleanup();
       },
-    });
-  }
-}
-
-/**
- * A ReadableStream that only emits a single chunk.
- */
-class SingleChunkReadableStream extends ReadableStream<any> {
-  constructor(content: string = "", toolCalls?: any) {
-    super({
-      start(controller) {
-        const chunk: ChatCompletionChunk = {
-          choices: [
-            {
-              delta: {
-                role: "assistant",
-                content,
-                ...(toolCalls ? { tool_calls: toolCalls } : {}),
-              },
-            },
-          ],
-        };
-        writeChatCompletionChunk(controller, chunk);
-        writeChatCompletionEnd(controller);
-
-        controller.close();
-      },
-      cancel() {},
     });
   }
 }