+++ CopilotKit/packages/react-core/src/utils/fetch-chat-completion.ts
@@ -58,6 +58,7 @@ export async function fetchChatCompletion({
       ...copilotConfig.backendOnlyProps,
       ...excludeBackendOnlyProps(copilotConfig),
       ...(body ? { ...body } : {}),
+      ...(copilotConfig.cloud ? { cloud: copilotConfig.cloud } : {}),
     }),
     signal,
   });